requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusers and Transformers
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0

# Computer Vision
# Using headless version for server/headless environments
opencv-python-headless>=4.8.0
Pillow>=9.5.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0

# Data Loading and Processing
hdf5storage>=0.1.18
h5py>=3.8.0

# Progress Bars and Utilities
tqdm>=4.65.0

# Logging and Experiment Tracking
wandb>=0.15.0

# Additional Utilities
requests>=2.31.0
typing-extensions>=4.5.0